Bank Customer Churn project
"Bank Customer Churn Model: Predicting the Future of Finance! Unveiling the data-driven saga of customer churn in the banking world - From imbalanced data to hyperparameter tuning, this project sparks the journey of taking on predictive analytics and paving the way to customer loyalty. Stay tuned for the thrilling revelation of customer churn insights and the battle against imbalance!"